WiFi-Connected Temperature Monitoring System
This project involves the design and implementation of a WiFi-enabled temperature monitoring system using an ESP8266 microcontroller and a DHT22 temperature sensor. The system allows you to read temperature data from the DHT22 sensor and send it wirelessly to a remote server via HTTP, providing a simple solution for monitoring temperature remotely.

Components Used:
ESP8266 Microcontroller:

A WiFi-enabled microcontroller that enables communication with the internet. It is used here to connect the temperature sensor to the WiFi network and transmit the temperature data to a server.
DHT22 Temperature and Humidity Sensor:

A widely used sensor for measuring temperature and humidity. In this project, it is used solely to measure the temperature.
SoftwareSerial Library:

Allows communication with the ESP8266 module over software serial communication on pins 2 and 3 of the Arduino.
Beeceptor Mock Server:

A free HTTP mock server used to receive and log the temperature data sent by the ESP8266 module.